DurationStringHelper.kt takes care only about String, DateConverter takes care of date, hour formatting
