+++ app/src/main/java/pl/llp/aircasting/screens/dashboard/SessionViewMvcImpl.kt
@@ -20,7 +20,7 @@ import pl.llp.aircasting.screens.dashboard.charts.Chart
 import pl.llp.aircasting.screens.session_view.measurement_table_container.MeasurementsTableContainer
 import pl.llp.aircasting.screens.session_view.measurement_table_container.SessionCardMeasurementsTableContainer
 import kotlinx.android.synthetic.main.expanded_session_view.view.*
-import pl.llp.aircasting.lib.DateHelper
+import pl.llp.aircasting.lib.DurationStringHelper
 
 abstract class SessionViewMvcImpl<ListenerType>: BaseObservableViewMvc<ListenerType>,
     SessionViewMvc<ListenerType> {
@@ -190,7 +190,7 @@ abstract class SessionViewMvcImpl<ListenerType>: BaseObservableViewMvc<ListenerT
     protected fun bindSessionDetails() {
         val session = mSessionPresenter?.session
 
-        mDateTextView.text = DateHelper.createDurationString(session!!)
+        mDateTextView.text = DurationStringHelper.durationString(session!!)
         mNameTextView.text = session?.name
         mInfoTextView.text = session?.infoString()
     }
