+++ app/src/main/java/pl/llp/aircasting/screens/main/MainActivity.kt
@@ -46,7 +46,6 @@ class MainActivity: AppCompatActivity() {
 
         DatabaseProvider.setup(applicationContext)
         LocationHelper.setup(applicationContext)
-        DateHelper.setup(settings)
         DateConverter.setup(settings)
         Places.initialize(applicationContext, BuildConfig.PLACES_API_KEY)
